fix: 🐛 issue when notifyOnChangeOnly stops emitting at all
When a user provides notifyOnChangeOnly option, we are trying to compare
the new metadata with the old one by stringifying them. Turns out, that
JSON.stringify() does not work on Map, so it returns an empty object all
the time. Leading to broken comparsion that says that metadata never
changes.